requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==26.0
pillow==12.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypdfium2==5.3.0
python-dotenv==1.2.1
python-json-logger==4.0.0
PyYAML==6.0.3
regex==2026.1.15
safetensors==0.7.0
setuptools==70.2.0
shellingham==1.5.4
sympy==1.13.1
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
qdrant_client
ollama
nltk
sentence_transformers
langchain_text_splitters